BookAppAI/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
SQLAlchemy
asyncpg
pgvector
numpy
google-generativeai
psycopg2-binary
python-dotenv